areatrigger_involvedrelation.sql fixed for areatrigger_involvedrelation.sql usage
The make_full_db.sql was not adding the first SQL script (DROP TABLE) or
all of the second SQL script  (CREATE TABLE) of the
areatrigger_involvedrelation.sql file to the full_db.sql file. Only from
the PRIMARY KEY (`id`) part onwards was added to the full_db.sql file.

This fixes that.