full_db.sql refused to be applied to Two's world database
For all files that were not in the same format as Zero and One's full DB
files, when applying ful_db.sql, it would not copy over the entire
CREATE TABLE part of the script, only part of it.

Also, there wasn't a DROP TABLE in most of the files, to precede the
CREATE TABLE. Added that to all files.

Basically, Two's _fullDB folder's files are now in the same format as
One and Two's.

Tested and working under Linux and MS Windows.