Vyragosa / Time-lost proto-drake
Added missing spawns for both Vyragosa and Time-lost proto-drake.  Also added waypoints and pooling for the two.
Due to inconsistent and lack of data, much of this information has been piece together from several sources.
Spawn times and % rate within the pool are a work in progress and may change in the future.